signWithGpg should default to true
I was not aware of that gradle Property<Boolean> is set to false per
default, I figured it is not set. Moreover, bintray uses boolean which
is false per default as well. Hence we need to change the strategy.
We only consider our `sighWithGpg` and overwrite sign in project
evaluation.