move module-info.java logic into own plugin
The approach of creating a sourceSet in afterEvaluate for sub-projects
Is not stable enough. It worked in tests, it also worked for some
projects but for others it failed. Also, the implicit behaviour is
not ideal and applying kotlin.utils multiple times is not ideal either
because in such a case one has to set kotlinVersion multiple times.
Thus, introduce an own plugin.