implement local gpg signing - do not use bintray anymore
Bintray cannot be used for serious signing as they require to upload
the private key. Instead we sign via gradle sign plugin and upload the
signatures in the publication. In order that one does not have to define
gpgKeyRing etc. if one does not publish, we have to make sure we
validate those information only if we publish (e.g. to local or to
bintray)