re-add publish plugin again but without uploading, support new MPP
I first wanted to include io.github.gradle-nexus:publish-plugin into
the publish plugin but saw that there are problems with this plugin
regarding coupling with the used kotlin version on one hand and gradle
version 7.1.1 problems (I did not take a closer look) on the other.

So this plugin does everything needed for the publication as such (set
up the publication if necessary, augment manifest, sign etc.) but the
uploading part needs to be done by something else.